Move _prepare_download to network.download
