+++ tests/unit/test_operations_prepare.py
@@ -1,5 +1,4 @@
 import hashlib
-import logging
 import os
 import shutil
 import sys
@@ -16,7 +15,6 @@ from pip._internal.operations.prepare import (
     Downloader,
     _copy_source_tree,
     _download_http_url,
-    _prepare_download,
     parse_content_disposition,
     sanitize_content_filename,
     unpack_file_url,
@@ -192,36 +190,6 @@ def test_download_http_url__no_directory_traversal(tmpdir):
     # The file should be downloaded to download_dir.
     actual = os.listdir(download_dir)
     assert actual == ['out_dir_file']
-
-
-@pytest.mark.parametrize("url, headers, from_cache, expected", [
-    ('http://example.com/foo.tgz', {}, False,
-        "Downloading http://example.com/foo.tgz"),
-    ('http://example.com/foo.tgz', {'content-length': 2}, False,
-        "Downloading http://example.com/foo.tgz (2 bytes)"),
-    ('http://example.com/foo.tgz', {'content-length': 2}, True,
-        "Using cached http://example.com/foo.tgz (2 bytes)"),
-    ('https://files.pythonhosted.org/foo.tgz', {}, False,
-        "Downloading foo.tgz"),
-    ('https://files.pythonhosted.org/foo.tgz', {'content-length': 2}, False,
-        "Downloading foo.tgz (2 bytes)"),
-    ('https://files.pythonhosted.org/foo.tgz', {'content-length': 2}, True,
-        "Using cached foo.tgz"),
-])
-def test_prepare_download__log(caplog, url, headers, from_cache, expected):
-    caplog.set_level(logging.INFO)
-    resp = MockResponse(b'')
-    resp.url = url
-    resp.headers = headers
-    if from_cache:
-        resp.from_cache = from_cache
-    link = Link(url)
-    _prepare_download(resp, link, progress_bar="on")
-
-    assert len(caplog.records) == 1
-    record = caplog.records[0]
-    assert record.levelname == 'INFO'
-    assert expected in record.message
 
 
 @pytest.fixture
